Generate comments from schema modifications more efficiently
Overriding all the `create_table` / `alter_table` generator methods was a
rather brutal way of doing things; Jeremy pointed me in the direction of
using the already-collected data about the indexes, constraints, etc that
were being created, so it turns out to be a lot easier just to scan that
churn through everything.

I also got rid of all the SQL generator stuff, because now that I'm not
derping around with overriding all those methods, I don't need all the extra
complexity it introduced.

Finally, I modified the signature of the `comment_on` method, so that it
takes an unambiguous description of the object, regardless of the type.